Shreyash (#13)
* Store Entity

* Store Entity & Staff Entity

* Thymeleaf Implementation, Explicit Mapping, Store Entity & Staff Entity

* Thymeleaf Implementation, Explicit Mapping, Store Entity & Staff Entity

* StoreController updated

* entity mapping and store api correction

* staff api

* 26 March api tests

* staff api

---------

+++ src/main/java/com/iem/FilmRentalStore/repository/CustomerRepository.java
@@ -9,5 +9,5 @@ public interface CustomerRepository extends JpaRepository<Customer, Short> {
     List<Customer> findByFirstNameContainingIgnoreCase(String firstName);
     List<Customer> findByLastNameContainingIgnoreCase(String lastName);
     List<Customer> findByEmailContainingIgnoreCase(String email);
-    List<Customer> findByStoreId(Short storeId);
+    List<Customer> findByStoreId(Byte storeId);
 }
